+++ mediawiki/mediawiki.py
@@ -16,7 +16,7 @@ from .mediawikipage import (MediaWikiPage)
 from .utilities import (memoize)
 
 URL = 'https://github.com/barrust/mediawiki'
-VERSION = '0.3.16'
+VERSION = '0.3.17'
 
 
 class MediaWiki(object):
@@ -559,11 +559,12 @@ class MediaWiki(object):
         '''
         self._check_query(category, 'Category must be specified')
 
+        max_pull = 5000
         search_params = {
             'list': 'categorymembers',
             'cmprop': 'ids|title|type',
             'cmtype': ('page|subcat' if subcategories else 'page'),
-            'cmlimit': (results if results is not None else 5000),
+            'cmlimit': (results if results is not None else max_pull),
             'cmtitle': 'Category:{0}'.format(category)
         }
         pages = list()
@@ -588,14 +589,25 @@ class MediaWiki(object):
                         tmp = tmp[9:]
                     subcats.append(tmp)
 
-            if 'continue' not in raw_res or last_cont == raw_res['continue']:
+            cont = raw_res.get('query-continue', False)
+            if cont and 'categorymembers' in cont:
+                cont = cont['categorymembers']
+            else:
+                cont = raw_res.get('continue', False)
+
+            if cont is False or last_cont == cont:
                 break
 
-            returned_results = returned_results + current_pull
+            returned_results += current_pull
             if results is None or (results - returned_results > 0):
-                last_cont = raw_res['continue']
+                last_cont = cont
             else:
                 finished = True
+
+            if results is not None and results - returned_results < max_pull:
+                search_params['cmlimit'] = results - returned_results
+
+            # print(last_cont)
         # end while loop
 
         if subcategories:
@@ -691,8 +703,8 @@ class MediaWiki(object):
             raise ValueError(msg)
 
         if depth is not None and depth < 1:
-            msg = ("CategoryTree: Parameter 'depth' must None (for the full "
-                   "tree) be greater than 0")
+            msg = ("CategoryTree: Parameter 'depth' must be either None "
+                   "(for the full tree) or be greater than 0")
             raise ValueError(msg)
 
         results = dict()