+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Current
 
+### Version 0.3.17
+
+* Add fix to use the `query-continue` parameter to continue to pull category
+members [issue #39](https://github.com/barrust/mediawiki/issues/39)
+* Better handle large categorymember selections
+
+
 ### Version 0.3.16
 
 * Add ability to turn off caching completely